% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_flag-bool.R
\name{create_bool_feature_flag}
\alias{create_bool_feature_flag}
\title{Creates an instance of a bool feature flag
with the specified bool value.}
\usage{
create_bool_feature_flag(value)
}
\arguments{
\item{value}{single logical determining whether the
flag should be enabled}
}
\value{
feature flag object of the bool value
}
\description{
Creates an instance of a bool feature flag
with the specified bool value.
}
\examples{
{
  enabled_flag <- create_bool_feature_flag(TRUE)
  disabled_flag <- create_bool_feature_flag(FALSE)
}
}
