% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_flag-connect_group.R
\name{create_connect_group_feature_flag}
\alias{create_connect_group_feature_flag}
\title{Creates an instance of a connect feature flag
that is enabled for specific groups}
\usage{
create_connect_group_feature_flag(groups)
}
\arguments{
\item{groups}{groups for which the feature flag
should be enabled}
}
\value{
feature flag that is enabled for specific groups
}
\description{
Creates an instance of a connect feature flag
that is enabled for specific groups
}
\examples{
{
  connect_group_flag <- create_connect_group_feature_flag(groups = c("group1", "group2"))
}
}
