% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_flag-connect_user.R
\name{create_connect_user_feature_flag}
\alias{create_connect_user_feature_flag}
\title{Creates an instance of a connect feature flag
with the specified bool value.}
\usage{
create_connect_user_feature_flag(users)
}
\arguments{
\item{users}{users for which the feature flag
should be enabled}
}
\value{
feature flag that is enabled for specific users
}
\description{
Creates an instance of a connect feature flag
with the specified bool value.
}
\examples{
{
  connect_user_flag <- create_connect_user_feature_flag(users = c("user1", "user2"))
}
}
