% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_flag.R
\name{create_feature_flag}
\alias{create_feature_flag}
\title{Creates the base of a feature flag.}
\usage{
create_feature_flag()
}
\value{
instance of a base feature flag.
}
\description{
It should not be used directly, but only
as a prerequisite when creating concrete feature flag.
}
