% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_flag-percentage.R
\name{create_percentage_feature_flag}
\alias{create_percentage_feature_flag}
\title{Creates an instance of a percentage feature flag
with a specified chance of being enabled}
\usage{
create_percentage_feature_flag(percentage)
}
\arguments{
\item{percentage}{chance of being enabled e.g. 1 for always
being enabled}
}
\value{
feature flag object of the percentage type
}
\description{
Creates an instance of a percentage feature flag
with a specified chance of being enabled
}
\examples{
{
  always_enabled_flag <- create_percentage_feature_flag(percentage = 1)
  randomly_enabled_flag <- create_percentage_feature_flag(percentage = 0.5)
}
}
