% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{feature_if}
\alias{feature_if}
\title{Evaluates the provided expression if the feature flag
is enabled.}
\usage{
feature_if(feature_flag, expr)
}
\arguments{
\item{feature_flag}{flag which defines whether the provided
expression should be evaluated}

\item{expr}{expression to evalute when the feature_flag is enabled}
}
\description{
Evaluates the provided expression if the feature flag
is enabled.
}
