% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{feature_ifelse}
\alias{feature_ifelse}
\title{Evaluates one or the other expression based on whether the feature flag
is enabled.}
\usage{
feature_ifelse(feature_flag, true_expr, false_expr)
}
\arguments{
\item{feature_flag}{flag which defines which expression should be
evaluated}

\item{true_expr}{expression to evalute when the feature_flag is enabled}

\item{false_expr}{expression to evalute when the feature_flag is disabled}
}
\description{
Evaluates one or the other expression based on whether the feature flag
is enabled.
}
