% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_flag-percentage.R
\name{is_enabled.percentage_feature_flag}
\alias{is_enabled.percentage_feature_flag}
\title{Checks if the given percentage flag is enabled}
\usage{
\method{is_enabled}{percentage_feature_flag}(feature_flag)
}
\arguments{
\item{feature_flag}{flag to be checked whether it is
enabled}
}
\value{
TRUE if the feature flag is enabled.
}
\description{
Checks if the given percentage flag is enabled
}
\examples{
{
  enabled_flag <- create_percentage_feature_flag(1)

  if (is_enabled(enabled_flag)) {
    print("The flag is enabled!")
  }
}
}
