% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_flag.R
\name{is_enabled}
\alias{is_enabled}
\title{Checks if the given feature flag is enabled.}
\usage{
is_enabled(feature_flag)
}
\value{
TRUE if the feature flag is enabled.
}
\description{
Checks if the given feature flag is enabled.
}
